requirements.txt
python-telegram-bot==20.1
mysql-connector-python==8.0.33
